--- app/src/main/java/ua/sviatkuzbyt/cityweather/ui/elements/basic/screens/LoadScreen.kt
@@ -1,27 +1,0 @@
-package ua.sviatkuzbyt.cityweather.ui.elements.basic.screens
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.width
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace36
-
-@Composable
-fun LoadScreen(){
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ){
-        CircularProgressIndicator(
-            modifier = Modifier.width(sizeSpace36),
-            color = MaterialTheme.colorScheme.primary,
-            trackColor = MaterialTheme.colorScheme.surfaceContainer,
-            strokeWidth = 2.dp
-        )
-    }
-}
