+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/elements/other/saveableCoroutineCall.kt
@@ -1,4 +1,4 @@
-package ua.sviatkuzbyt.cityweather.ui.elements
+package ua.sviatkuzbyt.cityweather.ui.elements.other
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
