+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/elements/screens/Screen.kt
@@ -1,12 +1,9 @@
-package ua.sviatkuzbyt.cityweather.ui.elements.basic.screens
+package ua.sviatkuzbyt.cityweather.ui.elements.screens
 
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace16
 
 @Composable
 fun Screen(
