+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/elements/screens/ScreenPlugs.kt
@@ -1,14 +1,17 @@
-package ua.sviatkuzbyt.cityweather.ui.elements.basic.screens
+package ua.sviatkuzbyt.cityweather.ui.elements.screens
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyItemScope
+import androidx.compose.material3.CircularProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.TextBasicCenter
+import androidx.compose.ui.unit.dp
+import ua.sviatkuzbyt.cityweather.ui.elements.other.TextBasicCenter
 import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace16
 import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace36
 
@@ -23,7 +26,8 @@ fun LazyItemScope.TextPlugList(
     ){
         TextBasicCenter(
             text = text,
-            modifier = modifier
+            modifier = modifier,
+            color = MaterialTheme.colorScheme.onPrimaryContainer
         )
     }
 }
@@ -41,6 +45,21 @@ fun TextPlug(
             text = text,
             modifier = modifier,
             color = MaterialTheme.colorScheme.onPrimaryContainer
+        )
+    }
+}
+
+@Composable
+fun LoadPlug(){
+    Box(
+        modifier = Modifier.fillMaxSize(),
+        contentAlignment = Alignment.Center
+    ){
+        CircularProgressIndicator(
+            modifier = Modifier.width(sizeSpace36),
+            color = MaterialTheme.colorScheme.primary,
+            trackColor = MaterialTheme.colorScheme.surfaceContainer,
+            strokeWidth = 2.dp
         )
     }
 }
