+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/screens/cities/elements/AddCityDialog.kt
@@ -31,11 +31,11 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.text.input.KeyboardCapitalization
 import ua.sviatkuzbyt.cityweather.R
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.TextHead
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.SpacerMedium
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.SpacerSmall
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.TextBasic
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.shapeContainer
+import ua.sviatkuzbyt.cityweather.ui.elements.other.TextHead
+import ua.sviatkuzbyt.cityweather.ui.elements.other.SpacerMedium
+import ua.sviatkuzbyt.cityweather.ui.elements.other.SpacerSmall
+import ua.sviatkuzbyt.cityweather.ui.elements.other.TextBasic
+import ua.sviatkuzbyt.cityweather.ui.elements.other.shapeContainer
 import ua.sviatkuzbyt.cityweather.ui.theme.Red
 import ua.sviatkuzbyt.cityweather.ui.theme.WhiteTransparent
 import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace20
