+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/screens/cities/elements/list/elements/ControlMenu.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import ua.sviatkuzbyt.cityweather.R
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.TextBasic
+import ua.sviatkuzbyt.cityweather.ui.elements.other.TextBasic
 import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace8
 
 @Composable
