+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/screens/cities/elements/list/elements/DetailElements.kt
@@ -18,10 +18,10 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import ua.sviatkuzbyt.cityweather.data.structures.cities.CityItemColors
 import ua.sviatkuzbyt.cityweather.data.structures.cities.CityItemDetailData
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.SpacerSmall
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.TextBasic
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.TextBasicSmall
-import ua.sviatkuzbyt.cityweather.ui.elements.basic.elements.shapeContainer
+import ua.sviatkuzbyt.cityweather.ui.elements.other.SpacerSmall
+import ua.sviatkuzbyt.cityweather.ui.elements.other.TextBasic
+import ua.sviatkuzbyt.cityweather.ui.elements.other.TextBasicSmall
+import ua.sviatkuzbyt.cityweather.ui.elements.other.shapeContainer
 import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace20
 import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace48
 import ua.sviatkuzbyt.cityweather.ui.theme.sizeSpace8
